+++ src/Symfony/Component/Notifier/Bridge/RocketChat/RocketChatOptions.php
@@ -17,28 +17,36 @@ use Symfony\Component\Notifier\Message\MessageOptionsInterface;
  * @author Jeroen Spee <https://github.com/Jeroeny>
  *
  * @see https://rocket.chat/docs/administrator-guides/integrations/
+ * @see https://developer.rocket.chat/reference/api/rest-api/endpoints/core-endpoints/chat-endpoints/postmessage
  */
 final class RocketChatOptions implements MessageOptionsInterface
 {
     /** prefix with '@' for personal messages */
     private ?string $channel = null;
 
-    /** @var mixed[] */
+    /** @var string[]|string[][] */
     private array $attachments;
 
+    /** @var string[] */
+    private array $payload;
+
     /**
-     * @param string[] $attachments
+     * @param string[]|string[][] $attachments
+     * @param string[]            $payload
      */
-    public function __construct(array $attachments = [])
+    public function __construct(array $attachments = [], array $payload = [])
     {
         $this->attachments = $attachments;
+        $this->payload = $payload;
     }
 
     public function toArray(): array
     {
-        return [
-            'attachments' => [$this->attachments],
-        ];
+        if ($this->attachments) {
+            return $this->payload + ['attachments' => array_is_list($this->attachments) ? $this->attachments : [$this->attachments]];
+        }
+
+        return $this->payload;
     }
 
     public function getRecipientId(): ?string