+++ src/Symfony/Component/Notifier/Bridge/RocketChat/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+6.2
+---
+
+ * Add the ability to send multiple attachments in a message
+ * Add the ability to send custom payload data to Rocket.Chat webhooks
+
 5.3
 ---